PR review change Requestion
Typo in create_table.py, description of name as "Name of the new tabl". This should be corrected to "Name of the new table".
There is no validation of base_id in create_table.py, providing base_id as an empty string seems incorrect.
there is no newline at the end of newly added files. This is not a major issue but is a best practice for Unix-based systems.